gst verification template/requirements.txt
flask
requests
bs4
uvicorn
asgiref
Pillow
